backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
openai==1.30.0
anthropic==0.7.0
PyYAML==6.0.1
python-dotenv==1.0.0
fastmcp 